significant cleanup and test fixes
