+++ test/test_external.py
@@ -151,24 +151,28 @@ class TestLoadInterface(unittest.TestCase):
     def test_sentiment_model(self):
         interface_info = gr.external.load_interface("models/distilbert-base-uncased-finetuned-sst-2-english", alias="sentiment_classifier")
         io = gr.Interface(**interface_info)
+        io.api_mode = True
         output = io("I am happy, I love you.")
         self.assertGreater(output['Positive'], 0.5)
 
     def test_image_classification_model(self):
         interface_info = gr.external.load_interface("models/google/vit-base-patch16-224")
         io = gr.Interface(**interface_info)
+        io.api_mode = True
         output = io("test/test_data/lion.jpg")
         self.assertGreater(output['lion'], 0.5)
 
     def test_translation_model(self):
         interface_info = gr.external.load_interface("models/t5-base")
         io = gr.Interface(**interface_info)
+        io.api_mode = True
         output = io("My name is Sarah and I live in London")
         self.assertEquals(output, 'Mein Name ist Sarah und ich lebe in London')
 
     def test_numerical_to_label_space(self):
         interface_info = gr.external.load_interface("spaces/abidlabs/titanic-survival")
         io = gr.Interface(**interface_info)
+        io.api_mode = True
         output = io("male", 77, 10)
         self.assertLess(output['Survives'], 0.5)
 
@@ -179,6 +183,7 @@ class TestLoadInterface(unittest.TestCase):
 
         interface_info = gr.external.load_interface("spaces/abidlabs/image-identity")
         io = gr.Interface(**interface_info)
+        io.api_mode = True
         output = io("test/test_data/lion.jpg")
         assertIsFile(output)
