+++ test/test_interfaces.py
@@ -24,43 +24,6 @@ def captured_output():
         sys.stdout, sys.stderr = old_out, old_err
 
 class TestInterface(unittest.TestCase):
-    # send_error_analytics should probably actually be a method of Interface
-    # (so it doesn't have to take the 'enabled' argument)
-    # and since it's specific to the launch method, it should probably be
-    # renamed to send_launch_error_analytics.
-    # these tests test its current behavior
-    @mock.patch("requests.post")
-    def test_error_analytics_doesnt_crash_on_connection_error(self, mock_post):
-        mock_post.side_effect = requests.ConnectionError()
-        send_error_analytics(True)
-        mock_post.assert_called()
-        
-    @mock.patch("requests.post")
-    def test_error_analytics_doesnt_post_if_not_enabled(self, mock_post):
-        send_error_analytics(False)
-        mock_post.assert_not_called()
-     
-    @mock.patch("requests.post")   
-    def test_error_analytics_successful(self, mock_post):
-        send_error_analytics(True)
-        mock_post.assert_called()
-                
-    # as above, send_launch_analytics should probably be a method of Interface
-    @mock.patch("requests.post")
-    def test_launch_analytics_doesnt_crash_on_connection_error(self, mock_post):
-        mock_post.side_effect = requests.ConnectionError()
-        send_launch_analytics(analytics_enabled=True,
-                              inbrowser=True, is_colab="is_colab",
-	                              share="share", share_url="share_url")
-        mock_post.assert_called()
-        
-    @mock.patch("requests.post")
-    def test_launch_analytics_doesnt_post_if_not_enabled(self, mock_post):
-        send_launch_analytics(analytics_enabled=False,
-                              inbrowser=True, is_colab="is_colab",
-	                              share="share", share_url="share_url")
-        mock_post.assert_not_called()
-     
     def test_reset_all(self):
         interface = Interface(lambda input: None, "textbox", "label")
         interface.close = mock.MagicMock()
@@ -160,19 +123,6 @@ class TestInterface(unittest.TestCase):
         self.assertEqual(len(interface.examples[0]), 1)
         interface.close()
 
-    def test_launch_counter(self):
-        with tempfile.NamedTemporaryFile() as tmp:
-            with mock.patch('gradio.interface.JSON_PATH', tmp.name):
-                interface = Interface(lambda x: x, "textbox", "label")
-                os.remove(tmp.name)
-                interface.launch(prevent_thread_lock=True)
-                with open(tmp.name) as j:
-                    self.assertEqual(json.load(j)['launches'], 1)
-                interface.launch(prevent_thread_lock=True)
-                with open(tmp.name) as j:
-                    self.assertEqual(json.load(j)['launches'], 2)
-                interface.close()
-    
     @mock.patch('IPython.display.display')
     def test_inline_display(self, mock_display):
         interface = Interface(lambda x: x, "textbox", "label")
